requirements.txt
neat-python==0.92
pygame==2.5.2
tqdm==4.66.4
